fixed sorting of group phase in interface by tracking additional evaluation data
Before, the frontend would have no knowledge about the details of the group phase (only the wins and draws and thus the "score"). Now there is an additional value for the average group phase rank of each time, which should finally solve the sorting issues once and for all. (As they are now always in sync with the simulation.)